Overview - How it works page
First draft of how it works. Wanted to take a break but lock this in.
Images are being worked on to add throughout.
Links for other areas of the docs will be added throughout.
Suggestions on organizing the flow of this to make the most sense from start to finish
Suggestions welcome on removing any of this from an overview and having a subpage.
Suggestions on any other topic that should be included in an overview of the platform for general understanding.